Reorganize validations and mutators for clarity.  The core pipeline of the proxy should be far more obvious now, making it easier to land features like quotas.
