+++ lib/proxy/index.js
@@ -4,16 +4,7 @@ var connect = require('connect');
 var httpProxy = require('http-proxy');
 var util = require('util');
 
-var oauth_params = require('./oauth_params.js');
-var whitelist = require('./whitelist.js');
-var authenticator = require('./authenticator.js');
-var header_modifier = require('./header_modifier.js');
 var ProxyKeys = require('./keys.js');
-
-var apply_xforwarded_headers = header_modifier.applyXForwardedHeaders();
-var url_parser = authenticator.urlParser();
-
-var MAXIMUM_URL_LENGTH = 16*1024;
 
 // Increase the maxSockets managed by this process to ensure that we can keep up with many
 // concurrent connections under load.  Also, node-http-proxy uses this agent to manage
@@ -65,60 +56,63 @@ Proxy.prototype.start = function(cb) {
       return cb(err);
     }
 
-    var validate_whitelist = authenticator.whitelistValidator(this_obj);
-    var collect_oauth_params = oauth_params.collectOAuthParams;
-    var validate_oauth_params = oauth_params.oauthParamValidator(this_obj);
-    var validate_oauth_signature = authenticator.oauthValidator(this_obj);
-    var request_validator = authenticator.requestValidator(this_obj);
-    var modify_host_header = header_modifier.modifyHostHeaders(this_obj.config.from_port, this_obj.config.to_port);
+    // Validators
+    var oauth_param_sanity_validator = require('./validators/oauth_param_sanity_validator.js')(this_obj);
+    var oauth_signature_validator = require('./validators/oauth_signature_validator.js')(this_obj);
+    var oauth_timestamp_validator = require('./validators/oauth_timestamp_validator.js')(this_obj);
+    var quota_validator = require('./validators/quota_validator.js')(this_obj);
+    var request_sanity_validator = require('./validators/request_sanity_validator.js')(this_obj);
+    var url_length_validator = require('./validators/url_length_validator.js')(this_obj);
+    var whitelist_validator = require('./validators/whitelist_validator.js')(this_obj);
+    
+    // Mutators
+    var form_parser = connect.urlencoded();
+    var forward_header_mutator = require('./mutators/forward_header_mutator.js')(this_obj);
+    var host_header_mutator = require('./mutators/host_header_mutator.js')(this_obj);
+    var oauth_param_collector = require('./mutators/oauth_param_collector.js')(this_obj);
+    var query_string_parser = connect.query();
+    var restreamer = require('./mutators/restreamer.js')({stringify:require('querystring').stringify});
+    var url_parser = require('./mutators/url_parser.js')(this_obj);
 
-    var restreamer = require('connect-restreamer')({stringify:require('querystring').stringify});
     var proxy = httpProxy.createProxyServer({});
+
+    // The express server is wired up with a list of mutators and validators that are applied to
+    // each inbound request.
     this_obj.server = connect.createServer(
       // Test for minimum viable sanity for an inbound request.  Pass in the proxy object so that
       // the URI and Host header can be matched against the expected values, if provided.
-      request_validator,
+      request_sanity_validator,
+      // Reject request with URLs longer than 16kb
+      url_length_validator,
       // Unpack the body of POSTs so we can use them in signatures.  Note that this
       // will implicitly limit POST size to 1mb.  We may wish to add configuration around
       // this in the future.
-      connect.urlencoded(),
-      // Reject request with URLs longer than 16kb
-      function(req, res, next) {
-        if (req.originalUrl.length < MAXIMUM_URL_LENGTH) {
-          return next();
-        }
-
-        res.writeHead(413, "URL exceeds maximum allowed length for oauth_reverse_proxy");
-        res.end();
-      },
+      form_parser,
       // Parse query string
-      connect.query(),
+      query_string_parser,
       // Parse url once so that it's available in a clean format for the oauth validator
       url_parser,
       // Gather the oauth params from the request
-      collect_oauth_params,
+      oauth_param_collector,
       // Modify the request headers to add x-forwarded-*
-      apply_xforwarded_headers,
+      forward_header_mutator,
       // Check the request against our path/verb whitelist
-      validate_whitelist,
+      whitelist_validator,
+      // Validate that the request is within quota
+      // quota_validator,
       // Validate that the oauth params pass a set of viability checks (existence, version, etc)
-      validate_oauth_params,
+      oauth_param_sanity_validator,
+      // Validate that the timestamp of the request is legal
+      oauth_timestamp_validator,
       // Perform the oauth signature validation
-      validate_oauth_signature,
+      oauth_signature_validator,
       // Update the host header
-      modify_host_header,
+      host_header_mutator,
       // Since connect messes with the input parameters and we want to pass them through
       // unadulterated to the target, we need to add restreamer to the chain.  But we only
       // need to do this if we're given a formencoded request.
-      function(req, res, next) {
-        if (req.headers && req.headers['content-type'] &&
-            req.headers['content-type'].indexOf('application/x-www-form-urlencoded') === 0) {
-          // Reconstitute form body only if necessary.
-          restreamer(req, res, next);
-        } else {
-          next();
-        }
-      },
+      restreamer,
+      // Whew.  After all of that, we're ready to proxy the request.
       function(req, res) {
         // Proxy a web request to the target port on localhost using the provided agent.
         // If no agent is provided, node-http-proxy will return a connection: close.
