+++ lib/proxy/mutators/forward_header_mutator.js
@@ -1,5 +1,3 @@
-var uuid = require('node-uuid');
-
 // Return the port, either by parsing the host header or by determining whether oauth_reverse_proxy did ssl
 // termination for the request.
 function getPortForRequest(req) {
@@ -29,7 +27,7 @@ var DEFAULT_VIA_HEADER = '1.1 localhost (oauth_reverse_proxy v' + OAUTH_REVERSE_
 // Append x-forwarded-for, x-forwarded-port, x-forwarded-proto, x-forwarded-host, and via to the request
 // headers before proxying to the target server.  If these headers are already set, append information
 // received by oauth_reverse_proxy to the existing headers.
-exports.applyXForwardedHeaders = function() {
+module.exports = function() {
   return function(req, res, next) {
 
     var values = {
@@ -45,26 +43,6 @@ exports.applyXForwardedHeaders = function() {
 
     // Set or append a via header.
     safelySetHeader(req, 'via', DEFAULT_VIA_HEADER);
-
-    next();
-  };
-};
-
-// Adjust the host header to reflect the proxy port.
-exports.modifyHostHeaders = function(from_port, to_port) {
-
-  // The new host header should take the form :to_port unless to_port is 443 or 80.
-  var new_port = (to_port === 443 || to_port === 80) ? '' : ':' + to_port;
-
-  return function(req, res, next) {
-    var host = req.headers['host'];
-
-    // If the existing host header includes a port, replace it.  Otherwise, append the value of new_port
-    // to the existing host header.
-    var idx = host.indexOf(':');
-    req.headers['host'] = idx !== -1 ?
-      host.substring(0, idx) + new_port :
-      host + new_port;
 
     next();
   };
