+++ lib/proxy/validators/oauth_signature_validator.js
@@ -1,19 +1,16 @@
 var util = require('util');
 
-var sprintf = require('../sprintf.js').sprintf;
-var encoding = require('../encoding.js');
+var sprintf = require('../../sprintf.js').sprintf;
+var encoding = require('../../encoding.js');
 
-var badrequest = require('./messages.js').badrequest;
-var unauthorized = require('./messages.js').unauthorized;
+var unauthorized = require('../messages/unauthorized.js');
 
 var crypto = require('crypto');
 
-var Whitelist = require('./whitelist.js');
+var OAUTH_CONSUMER_KEY = require('../oauth_constants.js').OAUTH_CONSUMER_KEY;
+var OAUTH_SIGNATURE = require('../oauth_constants.js').OAUTH_SIGNATURE;
 
-var OAUTH_CONSUMER_KEY = require('./oauth_params.js').OAUTH_CONSUMER_KEY;
-var OAUTH_SIGNATURE = require('./oauth_params.js').OAUTH_SIGNATURE;
-
-var CONSUMER_KEY_HEADER = exports.CONSUMER_KEY_HEADER = 'x-oauth-reverse-proxy-consumer-key';
+var CONSUMER_KEY_HEADER = require('../oauth_constants.js').CONSUMER_KEY_HEADER;
 
 /**
  * Sorts the encoded key value pairs by encoded name, then encoded value
@@ -86,77 +83,10 @@ function constructStringsToSign(req) {
 
   return signable_strings;
 };
-
-/**
- * Parse the input url once and cache the parsed value in the request object.
- */
-exports.urlParser = function() {
-  return function(req, res, next) {
-    req.parsed_url = require('url').parse(req.url, false);
-    next();
-  };
-};
-
-/**
- * Create a whitelist validator.  If the request passes the whitelist, it is automatically proxied through.
- */
-exports.whitelistValidator = function(proxy){
-
-    var whitelist = new Whitelist(proxy.config.whitelist);
-
-    return function(req,res,next) {
-        req.whitelist_passed = whitelist.applyWhitelist(req);
-
-        if(req.whitelist_passed) {
-            proxy.logger.info("Proxying URL %s %s%s WHITELIST", req.method, req.headers.host, req.url);
-        }
-
-        return next();
-    };
-};
-
-/**
- * Create a request sanity validator.  If the request does not pass these tests, it is a failed request and
- * shouldn't make it to any of the other phases of validation.
- */
-exports.requestValidator = function(proxy) {
-
-  var keys = proxy.keys;
-
-  return function(req, res, next) {
-    if (!req || !req.headers || !req.method || !req.url) {
-      return badrequest(proxy.logger, req, res, 'Invalid request');
-    }
-
-    var check_collection = function(coll, comparison_string, error_message) {
-      for (var i=0; i<coll.length; ++i) {
-        var val = coll[i];
-        if (comparison_string.indexOf(val) !== -1) {
-          return true;
-        }
-      }
-
-      return unauthorized(proxy.logger, req, res, error_message);
-    };
-
-    /* istanbul ignore else */
-    if (proxy.config.required_uris && !check_collection(proxy.config.required_uris, req.url, 'Unmatched URI')) {
-      return false;
-    }
-
-    /* istanbul ignore else */
-    if (proxy.config.required_hosts && !check_collection(proxy.config.required_hosts, req.headers.host, 'Unmatched Host header')) {
-      return false;
-    }
-
-    next();
-  };
-};
-
 /**
  * Create an oauth validator using the provided keys.
  */
-exports.oauthValidator = function(proxy) {
+module.exports = function(proxy) {
 
   var keys = proxy.keys;
