+++ lib/proxy/validators/quota_validator.js
@@ -2,9 +2,9 @@
  * Create a quota validator.  If the request does not exceed our requests per second threshold, allow the request
  * through.  If it does, fail with a 403 error.
  */
-exports.quotaValidator = function(proxy){
+module.exports = function(proxy){
 
-  var quota = new proxy.config.quota;
+  var quota = proxy.config.quota;
 
   // If the quota has an interval, convert it from seconds to ms.
   var clear_interval = quota.interval;
