+++ lib/proxy/validators/url_length_validator.js
@@ -1,0 +1,12 @@
+var MAXIMUM_URL_LENGTH = require('../oauth_constants.js').MAXIMUM_URL_LENGTH;
+
+module.exports = function() {
+  return function(req, res, next) {
+    if (req.originalUrl.length < MAXIMUM_URL_LENGTH) {
+      return next();
+    }
+
+    res.writeHead(413, "URL exceeds maximum allowed length for oauth_reverse_proxy");
+    res.end();
+  };
+};
